[#98] Move methods to find commands into a separate module
