+++ iceoryx2-cli/iox2/Cargo.toml
@@ -11,3 +11,5 @@ rust-version = { workspace = true }
 version = { workspace = true }
 
 [dependencies]
+human-panic = "1.2.3"
+better-panic = "0.3.0"
